requirements.txt
langchain>=0.2.0
langchain-google-genai>=1.0.3
langchain-community>=0.2.0
chromadb>=0.5.0
tiktoken
python-dotenv
gradio>=4.0.0
pypdf
